port: apply akka-management PRs #1150-#1155 to pekko
- PR #1150: Already applied (fix flaky RU test script)
- PR #1151: Fix AppVersionRevision docs wording in rolling-updates.md
- PR #1152: Fix cyclic initialization in ClusterBootstrap, AppVersionRevision, PodDeletionCost
- PR #1153: Add KUBERNETES_NAMESPACE env var support in reference.conf + docs
- PR #1155: Ensure versionPromise always completes on failure + refactor AppVersionRevisionSpec into KubernetesApiSpec

Agent-Logs-Url: https://github.com/pjfanning/incubator-pekko-management/sessions/bdef1ab5-2a1c-4728-927c-8c89aa47597a
