+++ discovery-kubernetes-api/src/main/resources/reference.conf
@@ -29,6 +29,7 @@ pekko.discovery {
     #
     # Set this value to a specific string to override discovering the namespace using pod-namespace-path.
     pod-namespace = "<pod-namespace>"
+    pod-namespace = ${?KUBERNETES_NAMESPACE}
 
     # Domain of the k8s cluster
     pod-domain = "cluster.local"
