+++ management-cluster-bootstrap/src/main/scala/org/apache/pekko/management/cluster/bootstrap/ClusterBootstrap.scala
@@ -88,7 +88,7 @@ final class ClusterBootstrap(implicit system: ExtendedActorSystem) extends Exten
       }
       try {
         PekkoManagement(system).start().failed.foreach(autostartFailed)
-        ClusterBootstrap(system).start()
+        start()
       } catch {
         case NonFatal(ex) => autostartFailed(ex)
       }
